fix(hover): incorrect position with border. Fixes #1122
The border was not being factored in when deciding how to anchor the
signature window, causing it to sometimes overlap the current line.

Similarly, the border was not being considered when drawing the
signature window above teh current line, causing it to be offset two
lines away from the current line.